azure-functions/requirements.txt
azure-functions==1.23.0
azure-storage-blob==12.26.0
azure-iot-hub==2.7.0
requests==2.31.0
Pillow==10.2.0

